refactor(google-calendar): delete vestigial gapi module + trim ambient types (#438)
`src/lib/google-calendar.ts` (578 lines) was the browser-side Google Identity
Services / gapi integration. PR #27 moved Google Calendar reads to the server
routes under `src/app/api/calendar/**`, leaving the runtime exports
(`fetchCalendarEvents`, `signInToGoogle`, `initGoogleAPI`, etc.) with no
production callers. PR #463 / phase 1 migrated the file's type re-exports off
the shim. This commit drops the module.

Knock-on changes:

- `src/types/gapi.d.ts` (386 → 269 lines) — drop the now-orphaned
  `google.accounts.oauth2` namespace and the duplicate `gapi` / `gapi.client`
  loader-namespace declarations (the latter overlap what `@types/gapi`
  already provides). The file now only augments `gapi.client.calendar` for
  the route-boundary cast in `src/app/api/calendar/events/**`. The header
  is rewritten to reflect that narrower scope and to point at #403 for the
  eventual full retirement.
- `src/lib/calendar-storage.ts` — inline the `GoogleCalendarAccount`
  interface (used only by that file's account-persistence helpers, which are
  themselves dead — tracked in #465 as a follow-up sweep). Minimal change
  required to keep type-check green after deleting `google-calendar.ts`.

Phase 2 of 3 for #438. Type-check clean, 2708 tests pass, lint reports 0 new
errors (5 pre-existing warnings on unrelated files).
